requirements.txt
amadeus==9.0.0
amqp==5.1.1
async-timeout==4.0.3
billiard==4.1.0
blinker==1.6.2
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
numpy==1.26.0
pandas==2.1.1
prompt-toolkit==3.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.8
Werkzeug==2.3.7
